+++ components/Tooltip/actions.tsx
@@ -1,0 +1,9 @@
+export const TooltipAction = {
+  HIDE: 'hide',
+  UPDATE: 'updatePosition',
+  START: 'startUpdate',
+  FINISH: 'finishUpdate',
+  FAIL: 'updateFailed'
+};
+
+export default TooltipAction;
